refactor: renamed a bunch of stuff to zsil
