+++ .gitignore
@@ -6,6 +6,6 @@
 .idea/vcs.xml
 *.pyc
 *.iml
-ZachsStupidImageLibrary/ffWACK.png
-ZachsStupidImageLibrary/fWACK.png
-ZachsStupidImageLibrary/WACK.png
+zsil/ffWACK.png
+zsil/fWACK.png
+zsil/WACK.png
